+++ test/mainnet-fork/VaultBaseTest.t.sol
@@ -31,14 +31,12 @@ import {VaultConstants} from "contracts/lib/VaultConstants.sol";
 import {PendingAddress, PendingUint192} from "contracts/lib/PendingLib.sol";
 import "contracts/storage/TermMaxStorage.sol";
 
-abstract contract VaultBaseTest is ForkBaseTest { 
-
-    struct VaultTestRes{
+abstract contract VaultBaseTest is ForkBaseTest {
+    struct VaultTestRes {
         uint256 blockNumber;
         uint256 orderInitialAmount;
         MarketInitialParams marketInitialParams;
         OrderConfig orderConfig;
-
         TermMaxMarket market;
         IMintableERC20 ft;
         IMintableERC20 xt;
@@ -51,11 +49,8 @@ abstract contract VaultBaseTest is ForkBaseTest {
         ITermMaxOrder order;
         ITermMaxVault vault;
         VaultInitialParams vaultInitialParams;
-
-        uint currentTime;
-
+        uint256 currentTime;
         uint256 maxCapacity;
-
         address maker;
     }
 
@@ -78,9 +73,12 @@ abstract contract VaultBaseTest is ForkBaseTest {
         res.collateralPriceFeed = deployMockPriceFeed(res.marketInitialParams.admin);
         res.debtPriceFeed = deployMockPriceFeed(res.marketInitialParams.admin);
         res.oracle.setOracle(
-            address(res.marketInitialParams.collateral), IOracle.Oracle(res.collateralPriceFeed, res.collateralPriceFeed, 365 days)
+            address(res.marketInitialParams.collateral),
+            IOracle.Oracle(res.collateralPriceFeed, res.collateralPriceFeed, 365 days)
         );
-        res.oracle.setOracle(address(res.marketInitialParams.debtToken), IOracle.Oracle(res.debtPriceFeed, res.debtPriceFeed, 365 days));
+        res.oracle.setOracle(
+            address(res.marketInitialParams.debtToken), IOracle.Oracle(res.debtPriceFeed, res.debtPriceFeed, 365 days)
+        );
 
         res.marketInitialParams.marketConfig.maturity += uint64(block.timestamp);
         res.marketInitialParams.loanConfig.oracle = res.oracle;
@@ -97,9 +95,15 @@ abstract contract VaultBaseTest is ForkBaseTest {
 
         // set all price as 1 USD = 1e8 tokens
         uint8 debtDecimals = res.debtToken.decimals();
-        _setPriceFeedInTokenDecimal8(res.debtPriceFeed, debtDecimals, MockPriceFeed.RoundData(1, 1e8, block.timestamp, block.timestamp, 0));
+        _setPriceFeedInTokenDecimal8(
+            res.debtPriceFeed, debtDecimals, MockPriceFeed.RoundData(1, 1e8, block.timestamp, block.timestamp, 0)
+        );
         uint8 collateralDecimals = res.collateral.decimals();
-        _setPriceFeedInTokenDecimal8(res.collateralPriceFeed, collateralDecimals, MockPriceFeed.RoundData(1, 1e8, block.timestamp, block.timestamp, 0));
+        _setPriceFeedInTokenDecimal8(
+            res.collateralPriceFeed,
+            collateralDecimals,
+            MockPriceFeed.RoundData(1, 1e8, block.timestamp, block.timestamp, 0)
+        );
 
         res.vault = ITermMaxVault(deployVaultFactory().createVault(res.vaultInitialParams, 0));
 
@@ -115,7 +119,9 @@ abstract contract VaultBaseTest is ForkBaseTest {
         res.debtToken.approve(address(res.vault), res.orderInitialAmount);
         res.vault.deposit(res.orderInitialAmount, res.marketInitialParams.admin);
 
-        res.order = res.vault.createOrder(res.market, res.vaultInitialParams.maxCapacity, res.orderInitialAmount, res.orderConfig.curveCuts);
+        res.order = res.vault.createOrder(
+            res.market, res.vaultInitialParams.maxCapacity, res.orderInitialAmount, res.orderConfig.curveCuts
+        );
 
         vm.stopPrank();
 
@@ -134,7 +140,7 @@ abstract contract VaultBaseTest is ForkBaseTest {
     }
 
     function _testDeposit(VaultTestRes memory res) internal {
-        _buyXt(res, 48.219178e8, uint128(res.orderInitialAmount/100));
+        _buyXt(res, 48.219178e8, uint128(res.orderInitialAmount / 100));
         vm.warp(res.currentTime + 2 days);
         address lper2 = vm.randomAddress();
         uint256 amount2 = 20000e8;
@@ -151,7 +157,7 @@ abstract contract VaultBaseTest is ForkBaseTest {
 
     function _testRedeem(VaultTestRes memory res) internal {
         vm.warp(res.currentTime + 2 days);
-        _buyXt(res, 48.219178e8, uint128(res.orderInitialAmount/100));
+        _buyXt(res, 48.219178e8, uint128(res.orderInitialAmount / 100));
         vm.warp(res.currentTime + 4 days);
         address lper2 = vm.randomAddress();
         uint256 amount2 = 10000e8;
@@ -173,7 +179,7 @@ abstract contract VaultBaseTest is ForkBaseTest {
 
     function _testBadDebt(VaultTestRes memory res) internal {
         vm.warp(res.currentTime + 2 days);
-        _buyXt(res, 48.219178e8, uint128(res.orderInitialAmount/100));
+        _buyXt(res, 48.219178e8, uint128(res.orderInitialAmount / 100));
 
         vm.warp(res.currentTime + 3 days);
         address lper2 = vm.randomAddress();
@@ -188,10 +194,10 @@ abstract contract VaultBaseTest is ForkBaseTest {
         address borrower = vm.randomAddress();
         vm.startPrank(borrower);
         deal(borrower, 1e18);
-        uint collateralAmt = 1e18;
+        uint256 collateralAmt = 1e18;
         deal(address(res.collateral), borrower, collateralAmt);
         res.collateral.approve(address(res.gt), collateralAmt);
-        res.market.issueFt(borrower, uint128(res.orderInitialAmount/10), abi.encode(collateralAmt));
+        res.market.issueFt(borrower, uint128(res.orderInitialAmount / 10), abi.encode(collateralAmt));
         vm.stopPrank();
 
         vm.warp(res.currentTime + 92 days);
@@ -212,12 +218,13 @@ abstract contract VaultBaseTest is ForkBaseTest {
 
         uint256 shareToDealBadDebt = res.vault.balanceOf(lper2);
         uint256 withdrawAmt = res.vault.previewRedeem(shareToDealBadDebt);
-        if(withdrawAmt > badDebt) {
+        if (withdrawAmt > badDebt) {
             withdrawAmt = badDebt;
         }
 
         vm.startPrank(lper2);
-        (uint256 shares, uint256 collateralOut) = res.vault.dealBadDebt(address(res.collateral), withdrawAmt, lper2, lper2);
+        (uint256 shares, uint256 collateralOut) =
+            res.vault.dealBadDebt(address(res.collateral), withdrawAmt, lper2, lper2);
         assertEq(shares, shareToDealBadDebt);
         assertEq(collateralOut, (withdrawAmt * delivered) / badDebt);
         assertEq(res.vault.badDebtMapping(address(res.collateral)), badDebt - withdrawAmt);
@@ -234,5 +241,4 @@ abstract contract VaultBaseTest is ForkBaseTest {
         res.order.swapExactTokenToToken(res.debtToken, res.xt, taker, tokenAmtIn, xtAmtOut);
         vm.stopPrank();
     }
-
 }
